+++ tests/unit/test_mongodb_provider.py
@@ -27,8 +27,9 @@ DEPARTED_IDS = [None, 0]
 
 @pytest.fixture(autouse=True)
 def patch_upgrades(monkeypatch):
-    monkeypatch.setattr("charm.k8s_upgrade._Partition.get", lambda *args, **kwargs: 0)
-    monkeypatch.setattr("charm.k8s_upgrade._Partition.set", lambda *args, **kwargs: None)
+    monkeypatch.setattr("charms.mongodb.v0.upgrade_helpers.AbstractUpgrade.in_progress", False)
+    monkeypatch.setattr("charm.kubernetes_upgrades._Partition.get", lambda *args, **kwargs: 0)
+    monkeypatch.setattr("charm.kubernetes_upgrades._Partition.set", lambda *args, **kwargs: None)
 
 
 class TestMongoProvider(unittest.TestCase):
